saffron.api/pyproject.toml

[tool.poetry]
name = "saffron.api"
version = "0.1.0"
description = "Backend for saffron.api with FastAPI and MongoDB"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110"
uvicorn = {extras = ["standard"], version = "^0.29"}
motor = "^3.4.0"
pydantic = "^2.6"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1"
httpx = "^0.27"

python-jose = "^3.3"
